Fixed some small issues
