+++ packages/data-sheet/src/actions/toolbar.module.sass
@@ -4,14 +4,15 @@
   gap: 8px
   align-items: center
   min-height: 32px
+  padding: 4px 8px
+  background-color: var(--panel-secondary-background-color)
 
 // Leading label naming the current selection shape (no icon) — reads as the
 // toolbar's title.
 .toolbar-title
   font-weight: 600
-  font-size: 0.9em
-  opacity: 0.7
   white-space: nowrap
+  min-width: 100px
 
 .spacer
   flex-grow: 1
@@ -36,6 +37,12 @@
   gap: 4px
   margin-bottom: 4px
   align-items: center
+  justify-content: flex-start
+
+  &>.group
+    display: flex
+    flex-direction: row
+    gap: 4px
 
 // The current filter window shown after the filter name on its tag
 .filter-window
