+++ packages/data-sheet/src/components/index.ts
@@ -4,3 +4,4 @@ export * from "./colors";
 export * from "./text-areas";
 export * from "./actions";
 export { InfoBar, LoadProgressIndicator } from "./view-info.ts";
+export { ScrollToRowControl } from "./scroll-to-row.ts";
