+++ packages/data-sheet/src/index.ts
@@ -5,7 +5,8 @@ export * from "./provider";
 export * from "./utils";
 export * from "./types";
 export * from "./actions";
-export * from "./actions/column-filter";
 export * from "./filters/operators";
 
 export { Cell } from "@blueprintjs/table";
+export { CellRendererDebugOverlay } from "./utils";
+export { ScrollToRowControl } from "./components";
