+++ packages/data-sheet/src/utils/index.ts
@@ -1,6 +1,7 @@
 export * from "./column-spec";
 export * from "./key-handling";
 export * from "./validation";
+export { CellRendererDebugOverlay } from "./debug-overlay.ts";
 
 export function range(arr: number[]) {
   if (arr.length != 2) throw new Error("Range must have two elements");
